+++ lib/skeem/interpreter.rb
@@ -12,12 +12,28 @@ module Skeem
     def initialize()
       @runtime = Runtime.new(Environment.new)
       add_primitives(runtime)
+      add_standard(runtime)
     end
     
     def run(source)
       @parser ||= Parser.new
       @ptree = parser.parse(source)
       return @ptree.root.evaluate(runtime)
+    end
+    
+    private
+    
+    def add_standard(aRuntime)
+      std_pathname = File.dirname(__FILE__) + '/standard/base.skm'
+      load_lib(std_pathname)
+    end
+    
+    def load_lib(aPathname)
+      lib_source = nil
+      File.open(aPathname, 'r') do |lib| 
+        lib_source = lib.read 
+        run(lib_source)
+      end
     end
   end # class
 end # module